If size_t happens to be 64 bits, then this will fail miserably. The format for %.*G requires a (int) then a (double). If size_t is a 64 bit thing, then this will just break terribly. Fix it by casting ndigits to an int.
git-svn-id: https://svn.aros.org/svn/aros/trunk@13157 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/gcvt.c
@@ -49,7 +49,7 @@
 
 ******************************************************************************/
 {
-    sprintf (buf, "%.*G", ndigit, number);
+    sprintf (buf, "%.*G", (int)ndigit, number);
 
     return buf;
 } /* sprintf */
